Refactor entity and default templates into single files
