+++ packages/generators/src/service/templates/schema.typebox.tpl.ts
@@ -1,11 +1,23 @@
 import { generator, toFile, when } from '@feathershq/pinion'
-import { fileExists, renderSource } from '../../commons'
+import { fileExists, localTemplate, renderSource } from '../../commons'
 import { ServiceGeneratorContext } from '../index'
+
+const authFieldsTemplate = (authStrategies: string[]) =>
+  authStrategies
+    .map((name) =>
+      name === 'local'
+        ? `  email: Type.String(),
+  password: Type.Optional(Type.String())`
+        : `  ${name}Id: Type.Optional(Type.String())`
+    )
+    .join(',\n')
 
 const template = ({
   camelName,
   upperName,
   relative,
+  authStrategies,
+  isEntityService,
   type,
   cwd,
   lib
@@ -18,6 +30,7 @@ import { ObjectIdSchema } from '@feathersjs/typebox'`
     : ''
 }
 import type { Static } from '@feathersjs/typebox'
+${localTemplate(authStrategies, `import { passwordHash } from '@feathersjs/authentication-local'`)}
 
 import type { HookContext } from '${relative}/declarations'
 import { dataValidator, queryValidator } from '${relative}/${
@@ -27,21 +40,35 @@ import { dataValidator, queryValidator } from '${relative}/${
 // Main data model schema
 export const ${camelName}Schema = Type.Object({
     ${type === 'mongodb' ? '_id: ObjectIdSchema()' : 'id: Type.Number()'},
-    text: Type.String()
+    ${isEntityService ? authFieldsTemplate(authStrategies) : `text: Type.String()`}
   }, { $id: '${upperName}', additionalProperties: false })
 export type ${upperName} = Static<typeof ${camelName}Schema>
 export const ${camelName}Validator = getValidator(${camelName}Schema, dataValidator)
 export const ${camelName}Resolver = resolve<${upperName}, HookContext>({})
 
-export const ${camelName}ExternalResolver = resolve<${upperName}, HookContext>({})
+export const ${camelName}ExternalResolver = resolve<${upperName}, HookContext>({
+  ${localTemplate(
+    authStrategies,
+    `// The password should never be visible externally
+  password: async () => undefined`
+  )}  
+})
 
 // Schema for creating new entries
-export const ${camelName}DataSchema = Type.Pick(${camelName}Schema, ['text'], {
+export const ${camelName}DataSchema = Type.Pick(${camelName}Schema, [
+  ${
+    isEntityService
+      ? authStrategies.map((name) => (name === 'local' ? `'email', 'password'` : `'${name}Id'`)).join(', ')
+      : `'text'`
+  }
+], {
   $id: '${upperName}Data'
 })
 export type ${upperName}Data = Static<typeof ${camelName}DataSchema>
 export const ${camelName}DataValidator = getValidator(${camelName}DataSchema, dataValidator)
-export const ${camelName}DataResolver = resolve<${upperName}, HookContext>({})
+export const ${camelName}DataResolver = resolve<${upperName}, HookContext>({
+  ${localTemplate(authStrategies, `password: passwordHash({ strategy: 'local' })`)}
+})
 
 // Schema for updating existing entries
 export const ${camelName}PatchSchema = Type.Partial(${camelName}Schema, {
@@ -49,11 +76,17 @@ export const ${camelName}PatchSchema = Type.Partial(${camelName}Schema, {
 })
 export type ${upperName}Patch = Static<typeof ${camelName}PatchSchema>
 export const ${camelName}PatchValidator = getValidator(${camelName}PatchSchema, dataValidator)
-export const ${camelName}PatchResolver = resolve<${upperName}, HookContext>({})
+export const ${camelName}PatchResolver = resolve<${upperName}, HookContext>({
+  ${localTemplate(authStrategies, `password: passwordHash({ strategy: 'local' })`)}
+})
 
 // Schema for allowed query properties
 export const ${camelName}QueryProperties = Type.Pick(${camelName}Schema, [
-  '${type === 'mongodb' ? '_id' : 'id'}', 'text'
+  '${type === 'mongodb' ? '_id' : 'id'}', ${
+  isEntityService
+    ? authStrategies.map((name) => (name === 'local' ? `'email'` : `'${name}Id'`)).join(', ')
+    : `'text'`
+}
 ])
 export const ${camelName}QuerySchema = Type.Intersect([
   querySyntax(${camelName}QueryProperties),
@@ -62,7 +95,21 @@ export const ${camelName}QuerySchema = Type.Intersect([
 ], { additionalProperties: false })
 export type ${upperName}Query = Static<typeof ${camelName}QuerySchema>
 export const ${camelName}QueryValidator = getValidator(${camelName}QuerySchema, queryValidator)
-export const ${camelName}QueryResolver = resolve<${upperName}Query, HookContext>({})
+export const ${camelName}QueryResolver = resolve<${upperName}Query, HookContext>({
+  ${
+    isEntityService
+      ? `
+  // If there is a user (e.g. with authentication), they are only allowed to see their own data
+  ${type === 'mongodb' ? '_id' : 'id'}: async (value, user, context) => {
+    if (context.params.user) {
+      return context.params.user.${type === 'mongodb' ? '_id' : 'id'}
+    }
+
+    return value
+  }`
+      : ''
+  }
+})
 `
 
 export const generate = (ctx: ServiceGeneratorContext) =>
